refactor(mission): extract Mission.SessionOps (session/workspace/task CRUD + github bindings)
Continues the god-module decomposition. SessionOps owns session/workspace
read+simple CRUD, task getters, project_name_taken?, and workspace↔GitHub
bindings (exact original bodies preserved: unique-constraint changeset
errors + owner/repo ordering). Mission 6774->~6400. Transactional task
lifecycle stays in Mission.

mix precommit: 2252/0 (1 excluded), CI guard passed.
